# perf(ops): support `Cow<'_, str>`