Inline `String` and `Option<String>` args in the #[op] macro. Also implements deserialization of `Cow<'_, str>` from v8::String - it is backed by a global stack allocated buffer.

5% improvement in op_url_parse by switching `href: String` to `href: Cow<'_, str>`

```
# main

deno run -A --unstable cli/bench/url_parse.js 50 1000000
time 1076 ms rate 929368
time 1057 ms rate 946073
time 1047 ms rate 955109
time 1045 ms rate 956937
time 1043 ms rate 958772
time 1046 ms rate 956022
time 1049 ms rate 953288
```

```
# This patch

target/release/deno run -A --unstable cli/bench/url_parse.js 50 1000000
time 1020 ms rate 980392
time 1012 ms rate 988142
time 1004 ms rate 996015
time 1001 ms rate 999000
time 1000 ms rate 1000000
time 1000 ms rate 1000000
time 1000 ms rate 1000000
```